Upload image user

+++ libs/user_model.php
@@ -190,22 +190,23 @@ class user extends gen_model{
         return true;   
     }
     
-    public function load_image($file_image)
-    {
+   public function load_image($file_image)
+   {
         if($this->attributes[$this->table_descr['key']] == -1){
             $this->err_descr = "Error, you have to initialize user, which user wants change the image?";
             return false;
         }
         $path = IMG_USER.$this->type.'/'.$this->attributes['user'];
-        if (file_exists($path)){
-            unlink($path);           
+        if($this->attributes['image'] != DEFAULT_IMG){
+            if (file_exists($path)){
+               unlink($path);
+            )
         }           
         //creazione del file
         if (!move_uploaded_file($file_image['userfile']['temp_name'], $path)){
             $this->err_descr = 'ERROR: an erro occurred while uplloading file';            
             return false;            
-        }
-        
+        }    
     }
     
 }
